billing_market/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
mysqlclient==2.1.1
openpyxl==3.1.1
phonenumbers==8.13.6
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
